flask_ml_app/requirements.txt
tensorflow>=2.6.0
pillow
Flask>=2.0.0
gunicorn==19.6.0
numpy==1.19.5
requests
Werkzeug